requirements.txt
apache-libcloud==1.5.0
appdirs==1.4.0
astroid==1.4.9
autopep8==1.2.4
backports.functools-lru-cache==1.3
backports.ssl-match-hostname==3.5.0.1
configparser==3.5.0
isort==4.2.5
lazy-object-proxy==1.2.2
libvirt-python==2.5.0
mccabe==0.6.1
packaging==16.8
pep8==1.7.0
pycurl==7.43.0
pylint==1.6.5
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
wrapt==1.10.8
